Extract ContextItem
